Fix os.confstr(): the result type of the C function is size_t, not int

+++ Modules/posixmodule.c
@@ -9140,7 +9140,7 @@ posix_confstr(PyObject *self, PyObject *args)
     PyObject *result = NULL;
     int name;
     char buffer[255];
-    int len;
+    size_t len;
 
     if (!PyArg_ParseTuple(args, "O&:confstr", conv_confstr_confname, &name))
         return NULL;
@@ -9157,7 +9157,7 @@ posix_confstr(PyObject *self, PyObject *args)
         }
     }
 
-    if ((unsigned int)len >= sizeof(buffer)) {
+    if (len >= sizeof(buffer)) {
         char *buf = PyMem_Malloc(len);
         if (buf == NULL)
             return PyErr_NoMemory();
